+++ lib/src/theme/components/option.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/widgets.dart';
+import 'package:shadcn_ui/src/utils/extensions/order_policy.dart';
 
 @immutable
 class ShadOptionTheme {
@@ -7,6 +8,7 @@ class ShadOptionTheme {
     this.hoveredBackgroundColor,
     this.padding,
     this.radius,
+    this.orderPolicy,
   });
 
   final bool merge;
@@ -14,6 +16,9 @@ class ShadOptionTheme {
   final Color? hoveredBackgroundColor;
   final EdgeInsets? padding;
   final BorderRadius? radius;
+
+  /// {@macro ShadOption.orderPolicy}
+  final WidgetOrderPolicy? orderPolicy;
 
   static ShadOptionTheme lerp(
     ShadOptionTheme a,
@@ -27,6 +32,7 @@ class ShadOptionTheme {
           Color.lerp(a.hoveredBackgroundColor, b.hoveredBackgroundColor, t),
       padding: EdgeInsets.lerp(a.padding, b.padding, t),
       radius: BorderRadius.lerp(a.radius, b.radius, t),
+      orderPolicy: t < 0.5 ? a.orderPolicy : b.orderPolicy,
     );
   }
 
@@ -35,6 +41,7 @@ class ShadOptionTheme {
     Color? hoveredBackgroundColor,
     EdgeInsets? padding,
     BorderRadius? radius,
+    WidgetOrderPolicy? orderPolicy,
   }) {
     return ShadOptionTheme(
       merge: merge ?? this.merge,
@@ -42,6 +49,7 @@ class ShadOptionTheme {
           hoveredBackgroundColor ?? this.hoveredBackgroundColor,
       padding: padding ?? this.padding,
       radius: radius ?? this.radius,
+      orderPolicy: orderPolicy ?? this.orderPolicy,
     );
   }
 
@@ -52,6 +60,7 @@ class ShadOptionTheme {
       hoveredBackgroundColor: other.hoveredBackgroundColor,
       padding: other.padding,
       radius: other.radius,
+      orderPolicy: other.orderPolicy,
     );
   }
 
@@ -63,7 +72,8 @@ class ShadOptionTheme {
         other.merge == merge &&
         other.hoveredBackgroundColor == hoveredBackgroundColor &&
         other.padding == padding &&
-        other.radius == radius;
+        other.radius == radius &&
+        other.orderPolicy == orderPolicy;
   }
 
   @override
@@ -71,6 +81,7 @@ class ShadOptionTheme {
     return merge.hashCode ^
         hoveredBackgroundColor.hashCode ^
         padding.hashCode ^
-        radius.hashCode;
+        radius.hashCode ^
+        orderPolicy.hashCode;
   }
 }