+++ lib/src/theme/components/select.dart
@@ -27,7 +27,6 @@ class ShadSelectTheme {
     this.effects,
     this.shadows,
     this.filter,
-    this.optionsOrderPolicy,
   });
 
   final bool merge;
@@ -52,9 +51,6 @@ class ShadSelectTheme {
   /// {@macro popover.shadows}
   final List<BoxShadow>? shadows;
 
-  /// {@macro ShadOption.orderPolicy}
-  final WidgetOrderPolicy? optionsOrderPolicy;
-
   static ShadSelectTheme lerp(
     ShadSelectTheme a,
     ShadSelectTheme b,
@@ -78,7 +74,6 @@ class ShadSelectTheme {
       clearSearchOnClose: t < 0.5 ? a.clearSearchOnClose : b.clearSearchOnClose,
       effects: t < 0.5 ? a.effects : b.effects,
       shadows: t < 0.5 ? a.shadows : b.shadows,
-      optionsOrderPolicy: t < .5 ? a.optionsOrderPolicy : b.optionsOrderPolicy,
     );
   }
 
@@ -120,7 +115,6 @@ class ShadSelectTheme {
       effects: effects ?? this.effects,
       shadows: shadows ?? this.shadows,
       filter: filter ?? this.filter,
-      optionsOrderPolicy: optionsOrderPolicy ?? this.optionsOrderPolicy,
     );
   }
 
@@ -142,7 +136,6 @@ class ShadSelectTheme {
       effects: other.effects,
       shadows: other.shadows,
       filter: other.filter,
-      optionsOrderPolicy: other.optionsOrderPolicy,
     );
   }
 
@@ -165,8 +158,7 @@ class ShadSelectTheme {
         other.clearSearchOnClose == clearSearchOnClose &&
         other.effects == effects &&
         other.shadows == shadows &&
-        other.filter == filter &&
-        other.optionsOrderPolicy == optionsOrderPolicy;
+        other.filter == filter;
   }
 
   @override
@@ -185,7 +177,6 @@ class ShadSelectTheme {
         clearSearchOnClose.hashCode ^
         effects.hashCode ^
         shadows.hashCode ^
-        filter.hashCode ^
-        optionsOrderPolicy.hashCode;
+        filter.hashCode;
   }
 }